added form validation, reload features for create/delete event

+++ Rally-FrontEnd/src/app/Events/event/event.component.ts
@@ -81,11 +81,24 @@ export class EventComponent implements OnInit {
 // }
 
 
+// deleteEvent() {
+//   this.eventService.deleteEvent(this.id).subscribe(data => {
+//     console.log(data);
+//   })
+//   this.router.navigate(["/events"]);
+// }
+
 deleteEvent() {
-  this.eventService.deleteEvent(this.id).subscribe(data => {
-    console.log(data);
-  })
-  this.router.navigate(["/events"]);
+  if(confirm("Are you sure you want to delete this event?")) {
+    this.eventService.deleteEvent(this.id).subscribe(data => {
+      console.log(data);
+    })
+    this.router.navigate(["/events"])
+  .then(() => {
+    window.location.reload();
+  });
+  }
+ 
 }
 
   // verifyLoggedIn() {
